+++ changelog.md
@@ -1,3 +1,4 @@
 - fixed pulley animation on first pull
 - fixed cannon boat shoot range
 - fixed flags banner config item renderer
+- changed hourglass load method some